gpg: Don't start agent for root and handle agent started by systemd

+++ modules/gpg/init.zsh
@@ -10,6 +10,11 @@ if (( ! $+commands[gpg-agent] )); then
   return 1
 fi
 
+# Don't start a gpg-agent for root
+if [[ "$UID" == 0 ]]; then
+  return 1
+fi
+
 # Set the default paths to gpg-agent files.
 _gpg_agent_conf="${GNUPGHOME:-$HOME/.gnupg}/gpg-agent.conf"
 _gpg_agent_env="${TMPDIR:-/tmp}/gpg-agent.env.$UID"
@@ -18,7 +23,7 @@ _gpg_agent_env="${TMPDIR:-/tmp}/gpg-agent.env.$UID"
 source "$_gpg_agent_env" 2> /dev/null
 
 # Start gpg-agent if not started.
-if [[ -z "$GPG_AGENT_INFO" && ! -S "${GNUPGHOME:-$HOME/.gnupg}/S.gpg-agent" ]]; then
+if [[ -z "$GPG_AGENT_INFO" && ! -S "${GNUPGHOME:-$HOME/.gnupg}/S.gpg-agent" && ! -S "/run/user/${UID}/gnupg/S.gpg-agent" ]]; then
   # Start gpg-agent if not started.
   if ! ps -U "$LOGNAME" -o pid,ucomm | grep -q -- "${${${(s.:.)GPG_AGENT_INFO}[2]}:--1} gpg-agent"; then
     eval "$(gpg-agent --daemon | tee "$_gpg_agent_env")"
@@ -44,6 +49,9 @@ if grep '^enable-ssh-support' "$_gpg_agent_conf" &> /dev/null; then
     gpg-connect-agent UPDATESTARTUPTTY /bye >/dev/null
   }
   add-zsh-hook preexec _gpg-agent-update-tty
+
+  # FIXME: Remove this workaround once upstream cleanly handles gpg-agent via systemd
+  unset SSH_AGENT_PID && export SSH_AUTH_SOCK="/run/user/$UID/gnupg/S.gpg-agent.ssh"
 fi
 
 # Clean up.
